Update: - merge image improve process into split - change project structure
