+++ src/hx011_split.py
@@ -5,10 +5,10 @@ import logging
 import os
 import struct
 
-import numpy as np
 import PIL.Image as pilI
-# import skimage
-import scipy.misc
+import numpy as np
+
+from helpTool import HanZi, ImFilterPipeline
 
 # Declare Log system
 logger = logging.getLogger('[Learn Tensorflow] Chinese hand writing')
@@ -30,37 +30,20 @@ logger.addHandler(fh)
 logger.addHandler(ch)
 
 # Declare const value
-base_dir = '../'
-training_data_dir = base_dir + 'data/HWDB1.1trn_gnt'
-test_data_dir = base_dir + 'data/HWDB1.1tst_gnt'
+# TODO change to your custom path
+base_dir = '/myWork/999_tmp/ai/data/handwrite'
+training_data_dir = base_dir + '/HWDB1.1trn_gnt'
+test_data_dir = base_dir + '/HWDB1.1tst_gnt'
 load_data_amount = 3
 # 汉字字典 (key : val) = (汉字 : Hanzi Class)
 dc = dict()
 
-
-class HanZi:
-    """
-    字典内部对象
-    同一个汉字存储在同一个文件夹下面
-    """
-    def __init__(self,
-                 folder_name,
-                 index):
-        self._folder_name = folder_name
-        self._index = index
-
-    @property
-    def folder_name(self):
-        return self._folder_name
-
-    @property
-    def index(self):
-        return self._index
-
-    def get_and_increase_index(self):
-        rtn = self._index
-        self._index += 1
-        return rtn
+# Declare image improve
+cim = ImFilterPipeline()
+pip_settings = cim.pipeline
+pip_settings["rotated"] = 1
+# pip_settings["gaussianBlur"] = 1
+pip_settings["gaussianBlur"] = 1
 
 
 def get_training_file_list(train_dir='data',
@@ -130,27 +113,9 @@ def resize_and_normalize_image(img):
     :param img: image array
     :return: resized image array
     """
-    # 补方
-    pad_size = abs(img.shape[0]-img.shape[1]) // 2
-    if img.shape[0] < img.shape[1]:
-        pad_dims = ((pad_size, pad_size), (0, 0))
-    else:
-        pad_dims = ((0, 0), (pad_size, pad_size))
-    img = np.lib.pad(img, pad_dims, mode='constant', constant_values=255)
+    processed_image = cim.filter(img)
 
-    # 缩放
-    # python 3.7
-    # img = skimage.transform.resize(img, (64 - 4*2, 64 - 4*2), mode='constant', anti_aliasing=True)
-    # python 3.5
-    img = scipy.misc.imresize(img, (64 - 4*2, 64 - 4*2))
-
-    img = np.lib.pad(img, ((4, 4), (4, 4)), mode='constant', constant_values=255)
-    assert img.shape == (64, 64)
-
-    img = img.flatten()
-    # 像素值范围-1到1
-    img = (img - 128) / 128
-    return img
+    return processed_image
 
 
 def get_data(file_list, folder='png/'):
@@ -210,10 +175,10 @@ if __name__ == '__main__':
     training_data_files = get_training_file_list(training_data_dir, target_data_amount=load_data_amount)
     test_data_files = get_training_file_list(test_data_dir, target_data_amount=load_data_amount)
 
-    training_img_list, training_label_list = get_data(training_data_files, '../data/train/')
+    training_img_list, training_label_list = get_data(training_data_files, base_dir + '/data/train/')
     logger.info("Load {} images for training.".format(len(training_label_list)))
 
-    test_img_list, test_label_list = get_data(test_data_files, '../data/test/')
+    test_img_list, test_label_list = get_data(test_data_files, base_dir + '/data/test/')
     logger.info("Load {} images for test.".format(len(test_label_list)))
 
     logger.info("== Job finished ==")
